+++ tests/test_async_spying.py
@@ -1,6 +1,5 @@
 """Test spying functionality."""
 
-# pylint: disable=missing-docstring
 import re
 
 from chainmock import mocker